+++ Tests/ArgumentCompleters/TabExpansion2.v5.build.ps1
@@ -1,0 +1,4 @@
+
+Mdbc.ArgumentCompleters.ps1
+
+. .\Tasks.ps1