+++ Tests/ArgumentCompleters/TabExpansionPlusPlus.build.ps1
@@ -1,0 +1,5 @@
+
+Import-Module TabExpansionPlusPlus
+Mdbc.ArgumentCompleters.ps1
+
+. .\Tasks.ps1